fix(core): Note::_size to size_t + decompress entry phdr clear (R50-18)
Memory-lifecycle audit found no triggerable leak/double-free/overflow:
cleanup_decompress covers all allocs, fill_* payload copies fit the
allocate reserve, every decompress failure path calls cleanup. Two cheap
hardening items applied:

- Note::_size int -> size_t (theoretical int truncation on >2GB note; not
  reachable now due to GetFile 64MB cap). Fixed the %d format specifier
  surfaced by the type change.
- decompress entry now clears _phdrs for consistency with the capture
  functions (defense against same-object generate-then-decompress reuse;
  the budget check already fail-closed on stale phdrs).

Confirmed dead code (left as-is): Acore class, debugstr/_hexbuf (never
called; debugstr reads a[size()] which is the valid NUL terminator).
Reverted an allocate assert->NULL change since this build (-O0 -g, no
NDEBUG) already aborts cleanly on OOM and guarding 6 memcpy callers
wasn't worth it.

Roundtrip + multithread regression pass.

+++ src/core.h
@@ -185,7 +185,9 @@ private:
     Note();
     int _type;
     char* _data;
-    int _size;
+    // R50-18: int→size_t——note 长度理论可超 2GB（int 截断成负数，fwrite 转
+    // size_t 巨量越界）。当前受 GetFile 64MB 上限约束不可触发，改类型消除隐患。
+    size_t _size;
 
     friend class Coredump;
 };
